backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Elisa <elisahonoratolira@uc.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"
pytz = "^2024.1"
sqlparse = "^0.5.0"
django-cors-headers = "^4.3.1"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^10.4.0"
django-admin-tailwind = "^0.0.31"
python-decouple = "^3.8"
model-bakery = "^1.20.0"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
moviepy = "^2.1.1"
django = "^5.1.4"
firebase-admin = "^6.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
